+++ python/rmm/_lib/memory_resource_wrappers.hpp
@@ -121,8 +121,8 @@ class binning_memory_resource_wrapper : public device_memory_resource_wrapper {
   }
 
   binning_memory_resource_wrapper(std::shared_ptr<device_memory_resource_wrapper> upstream_mr,
-                                  int min_size_exponent,
-                                  int max_size_exponent)
+                                  int8_t min_size_exponent,
+                                  int8_t max_size_exponent)
     : upstream_mr(upstream_mr),
       mr(std::make_shared<rmm::mr::binning_memory_resource<rmm::mr::device_memory_resource>>(
         upstream_mr->get_mr().get(), min_size_exponent, max_size_exponent))
